Scale harvest durability and building progress with time.

+++ src/main/java/com/vast/component/Building.java
@@ -6,5 +6,5 @@ import com.artemis.annotations.PooledWeaver;
 @PooledWeaver
 public class Building extends Component {
 	public String type = "unspecified";
-	public int progress = 0;
+	public float progress = 0.0f;
 }
